code dump importing Anthony's research
Anthony did a lot of research on tracking thread stacks that's immediately
applicable to properly detecting stack pivots. This imports a lot of that in
and makes sure we compile, but we certainly aren't in a functional state.

Next steps will be to get things functional on this imported code's terms
so we can essentially replicate its findings, then adapting this code to what
we want our PoC to do, which will involve more of the processing & decision
making happening in eBPF, and using bpf_send_signal to kill violating
processes.