Flattened commit
This PoC is "done" and not expected to change in any meaningful way.
Flatten all commits to a single commit to remove irrelevant history
and make sure we aren't releasing things like internal URLs. The README
and other documentation should be enough, and nearly all of the commit
history messages were notes-to-self for that moment in time.