Add wake_up_new_task meat, properly track thread stacks.
In a simple test, this removes the "unknown stack" errors from a quick
boot-up run of firefox. However because of still mysterious-to-me
issues with the `struct task_struct` type introducing errors in the
generated rust skeleton, we have to omit sending events to userland
from the wake_up_new_task eBPF program. But we still do the bookkeeping
to track thread stacks.