refactor message parsing in userland, redo wake_up_new_task events
as-is, the wake_up_new_task events were giving us issues. So instead
we just reuse the type for tracking stack areas and directly emit
those types to userland as events, so we can observe new stack regions
as we see them 'created' by new threads.

Also, refactor some of the event message parsing, since the current design
is a lot of different perf buffers, which leads to different handlers
with lots of boilerplate.