Add beginnings of tests
We'll need tests that emulate being exploited by a ROP chain which uses a stack
pivot. This brings in a basic makefile for building tests (and shellcode if we
want that). We'll probably mostly use C and inline assembly for our tests
though, to save us having to manually craft ROP chains.