Rudimentary test is working!
now we have a stack pivot program which will be caught as a stack pivot and
killed. This is a good first step of demonstrating the concept.

We should make killing configurable, just in case running our PoC causes
intense problems on our system.

We should also refactor our ringbufs and data types. We don't want to be using
an expanding list of ringbufs and types for debug output, and would prefer to
have a single ringbuf/type for reporting detected stack pivots.