Anthony's PoC nearly cleany integrated, ready for significant overhaul
We have moved debug output (which used to be via program-specific ringbufs)
into normal printks, reserving ringbuf use for reporting stack pivot events.

We add a return code to `find_vma` to aid in analyzing our `check_stack_vma`
function, so we know if a vma lookup failed or not.

fix bug in stack checking in execve probe.

disable killing while we refine stack checking logic.

We may have a few issues with the existing `check_stack_vma` logic for detecting
a stack pivot. Currently it (maybe not always 'correctly') detects when a stack
pointer is in an unknown stack region. However it only considers a 'stack pivot'
to be when a sp is not 'backed' by a VMA, which I think is really a segfault.