Add test for incorrect thread stack tracking issue, refactor tests
We were having a somewhat mysterious problem with detecting stack
pivots which produced multiple false positives. This test is a simple
reproduction of the issue.

When a multithreaded process would create a new thread, the call to clone
would incorrectly observe a stack pivot, based on the tracked stack region
of that task.