big changes, stack pivot check v1, some refactoring, README update
With Anthony's PoC mostly integrated, we start work on implementing an explicit
stack pivot check. The earlier vma check was insufficient and mostly was useful
for verifying the thread stack tracking.

We do some light refactoring by necessity; data_t needs to be smaller to fit in
the eBPF stack. Much more remains to be done.

Our initial stack pivot detection exposed a bug that took some time to track
down. Initially observed as a false positive stack pivot in firefox, and in
opening a new tab in gnome-terminal, the issue was replicable buy a simple
multi-threaded program. The thread group leader would call clone, and this
would be detected as a stack pivot. However the earlier execve call wouldn't
cause a false positive stack pivot.

This appeared to be due to behavior of clone/fork and execve on a process's
virtual memory. fork starts off the new process with a copy of the parent
process's virtual memory, thus the stack region is identical (but CoW pages so
of course there's VMA isolation). However execve has the kernel load a new
binary image and start it in the current process's address space, thus
essentially overwriting/rearranging the existing mappings.  This means the
stack region is almost certainly relocated elsewhere based on the new process
image. However the old stack mapping is still tracked in our own map. Thus
subsequent checks for this task find the old stack mapping leftover from the
parent process, not the new stack mapping.

The solution that appears to work in light testing is to invalidate the tracked
stack region at the end of our execve eBPF program. It's valid for checking the
execve call, but afterwards won't be valid so should be removed. Falling back
to the `start_stack` field from the task's mm member (mm_struct) appears
sufficient for checking the thread group leader for stack pivots.

Things are still in a very messy state and need considerable refactoring/tests.

We also refactor tests a bit (to reuse the GET_SP macro) and add a simple test
program that just reports its rsp register in main.