Hacky Golang exception, large code cleanup, debug sys_write hook
We remove a big chunk of commented code just to keep things clean.

We add an initial attempt at having an exception for Golang based
on our test case. It needs deeper exploration to make sure it's accurate,
and may still be a good idea to report in case the user knows they
won't be using any Golang applications. Golang applications
likely won't be issues with ROP chains due to memory safety, but
depending how the exception is implemented, any exception to avoid
false positives will leave a hole to sneak a stack pivot through
if the attacker can control memory allocations/get lucky with
application behavior.

Also add a hook to sys_write to make it more likely to catch false
positives. It's much easier to have various threads calling write
to print to stdout, than to call clone/execve. This is only suitable for
development & testing, and would have significant overhead in production.