Add stack pivot checking to mmap & mprotect syscalls
We already check on execve, which should catch ROP chains using a
stack pivot to launch a new binary image. However, it's also possible
for staged shellcode to allocate executable memory for a second stage
to be loaded into, or for a full-fledged payload to be delivered
and run in memory rather than using execve to replace the binary image
in the process. This adds stack pivot checks to the mmap and mprotect
syscalls.

In the future it would be good to restrict these to only mmap/mprotect
calls which involve the executable permission, to reduce overhead for
benign mmap/mprotect calls.