Clarify README, add stack pivoting to mmap/mprotect test, reduce mmap/mprotect events
The README now is explicit about needing to install `rustfmt`. This
tool is necessary for generating the rust skeleton of eBPF objects
built using the libbpf-rs crate.

Add stack pivoting to the mmap/mprotect test. This way the mmap
and mprotect calls (which are now creating suspicious RWX pages)
will trigger stack pivot events. This test will now generate two
events for a stack pivot, where previously it generated none.

We add a heuristic to the mmap/mprotect eBPF programs to bail early
if the executable bit is not included in the requested page protection.
This is aimed at catching staged shellcode techniques that use the
first stage to allocate executable memory for the second stage, rather
than execve. This helps us avoid overhead for less interesting mmap
and mprotect calls which aren't producing executable memory.