Vastly increase assumed good stack region to reduce false positives
We got lots of false positives on Ubuntu 22.04 LTS, kernel 5.15.0-46-generic.
apt install, firefox, many more. It's unclear what the root issue is but it
seems that somehow VMA behavior has changed. Firefox 104.0 on Ubuntu 20.04 LTS
cause no false positives, but on Ubuntu 22.04 LTS it does.

False positives when launching Firefox had stack pointers sometimes very far
outside our VMA that we decide is the legitimate stack. Meanwhile these tracked
VMAs are of only a few sizes, the largest being the 8Mib soft stack limit.
However stracing programs exhibiting false positives to see if they increase
this limit, it appears that we have two limits; soft and hard. The soft limit
is 8Mib, but the default hard limit on both Ubuntu versions tested is
unlimited, which practically means however much memory is physically present.
It seems that when the soft limit is surpassed as the stack grows, it page
faults and transparently allocates more memory to the stack.

So we don't have a nice and easy stack limit size we can use for comparisons.
We can however experiment with what false positives we get and see how much
larger we can expand our notion of a legitimate stack before it's useless.

It seems that, just based on the firefox false positives, we can probably expect
the upper 3 bytes of the stack pointer to match the discovered VMA. This should at
least be able to distinguish heap from stack addresses, but is much weaker than
we would like.