Attempt to use more threads & heap to force mmap calls in kotlin test
I'm very unfamiliar with Kotlin+GraalVM+NativeImage. To try and
stress test our stack pivot detection in the case of coroutines
in Kotlin using GraalVM to build a native image, we'll just start with
making big arrays from an object in a coroutine and hope that triggers mmaps.

This could use much improvement, but is at least a start.