handle execve errors, overhauled (but still broken) stack pivot check
The stack pivot check is overhauled to be more precise and strict.
However due to this there are some false positives that we should
either 1) debug more closely to determine the root cause and how
to more closely follow stack behavior for those edge-cases or 2)
loosen our check to be less precise but more forgiving.

Importantly we only remove tracked stack regions on a successful
execve. execve will essentially totally redo a process's virtual
memory by loading a new binary. However only if it's successful.
In the event that execve fails for some reason, the process's
virtual memory is not changed, and we should keep our tracked
stack region for that process.