Messy fix for false positives from strict checking
There are cases observed when launching firefox where we have a tracked stack
region for the thread group leader, and prefer the tracked stack over the vma
provided by mm->start_stack, but the sp actually does reside with
mm->start_stack. We just loosen the check to allow for a thread group
leader stack pointer to reside in either region and be considered
legitimate.

In the future it may be cleaner to root-cause _why_ we are tracking a
stack region in this case and address that. For thread group leaders my
expectation would that we only are tracking a stack region for the edge-case of
mm->start_stack (the initial value of sp when the task is first being created)
being the beginning address of a VMA. In practice the stack will actually
reside in the vma _previous_ to that one. We currently handle this
edge-case by finding the VMA one word before this sp value, and adding that
to our tracked stacks map.