Begin transitioning to new stack pivot event type
partly convert clone and execve probes. I don't know if
we have a good stack pivot test for clone, but we can make one later.

We only partly convert because we want to focus on the eBPF source
right now, and changing the event type that's emitted to userspace will
require changes rust-side. Let's convert the actual guts of our
probes before we do this, so that switch-over can be easy, since
it'll have to happen for all probes at the same time.