Use macros to simplify BPF programs that are only for checking rsp
We have a number of BPF programs that are only to check for a stack pivot, so
they're essentially the same besides a few identifiers/strings.  We use macros
to make them identical to ease maintenance. It shouldn't be easy to have one
missing an important piece of code simply because they're repeated boilerplate.

We decompose this "one big function" macro into some smaller sub-macros because
mmap/mprotect have a heuristic to short-circuit exit early if the requested
protection isn't executable. This is in hopes of avoiding some performance
overhead on mmap/mprotect invocations which won't be useful for staged
shellcode. So these BPF programs are a bit more custom, but still use the
macros to share as much code as possible with the other BPF programs which
simply check RSP.