add leader/non-leader option to mmap_mprotect and stackpivot tests
This is to make sure we exercise (at least a little bit) our different
detection cases of having a thread group leader, or non-leader.

Note that the mmap_mprotect tests only hits the mmap/mprotect syscalls,
and the stackpivot test only hits the execve syscall. We may want to add
tests later on to hit other syscalls we want to watch, which may be nice
to do with a more generic stack pivot simulation harness. But for now,
this is fine since the checking logic is identical. Hitting more syscalls
is only necessary for a more production-oriented test which is making sure
we have coverage of all the syscalls we want to cover.