Add optional sigkilling to PoC, to kill offending processes
This adds the option to sigkill a process which has a detected
stack pivot event. We use a singleton map to hold the 0/1 value
for if we're sigkilling or not, and add the clap crate as dependency
to easily take this value on the command line as an optional flag
-s (for 'sigkill'). This will make testing easy, considering if we
run across any false positives it could cause _big_ problems for
a running system.