Add info on running on kubernetes, since we finally figured it out.
This took some effort and googling, mostly to get the docker registry usable by
the kubernetes cluster's nodes. Setting up a docker registry on the cluster
with the docker-registry helm chart was straightforward.  However
kubernetes/docker expects https for registries. This means generating
certs/keys for the registry, configuring the docker registry to use them, and
configuring worker nodes with the appropriate CA to verify connections to the
cluster-internal docker registry.

This is documented in the README, but in short I deployed the docker-registry
helm chart, generated a self-signed CA cert/key with SAN entries for its
internal cluster IP/domain name, upgraded the docker-registry deployment to use
these via a kubernetes secret, and deployed the cert to the worker nodes under
/etc/docker/certs.d so verification would work. Searching for each of these
individual steps should help fill in the gaps.