Change Golang user stack check based on Anthony's research
Basically, Golang user stacks live in the runtime heap,
which has arenas that are allocated based on hints to mmap.
This isn't a 100% guarantee of address ranges stacks will be,
but it's a great place to start and is way better than manually
adding ranges.

Basically this lets us do a bit test on some of the higher bytes
of the address, to check that they match the expected format based
on the hints golang uses to allocate heap arenas.