Build container to 'latest' tag
This is probably basically standard practice. Makes sure we can
push new images to our registry and `kubernetes rollout restart`
the deployment to re-start pods with new containers.

Pulling latest containers in kubernetes is handy for dev but bad
for production. For production we'll need to start building with version
tags and using those in kubernetes deployment configs.