Add SIGUSR1 handling to print event stats
This makes it easier to see how many different events we're
seeing without having to shut down the PoC. Though, this does
require we manually interact with the cluster. Currently
this can be done by using `kubectl exec` to get an interactive
bash shell in the PoC container, and running `kill -10 1` to send
SIGUSR1 to the running PoC (PID 1 of its own PID namespace).