Rewrite parts about Golang, add table of cluster event breakdown
Now that we've run the PoC with the new Golang check for a while,
on a cluster with some light Traefik+Nextcloud usage, we're a bit
more confident that Golang false positives are now rare.

K8s is written in Go, as well as Traefik, so this seems like a decent test. It
would be nice to have fuller testing done in in-production clusters, but that's
left as future research.