Update License, Add Kernel Compatibility Note
The eBPF code is now only GPL and accurately reflected in
the LICENSE string used during loading. Also, add a note on
kernel compatibility (5.11-6.0). The rb-trees we use to look
up VMAs were removed in 6.1 in favor of maple trees. The code
as it is will not work on kernel 6.1+ until we add something
to handle this. That is, both code using the maple trees, as
well as code to determine which to use based on kernel version.